+++ omnilib/prototypes/functions/icon.lua
