+++ omnilib/prototypes/functions/locale-hierarchy.lua
