+++ omnilib/prototypes/functions/locale.lua
